extract static-build, new modular build modes
