+++ src/write/build-modes/live-dev-server.js
@@ -1,0 +1,31 @@
+import {logInfo} from '../../util/cli.js';
+
+export function getCLIOptions() {
+  // Stub.
+  return {};
+}
+
+export async function go({
+  _cliOptions,
+  _dataPath,
+  _mediaPath,
+  _queueSize,
+
+  _defaultLanguage,
+  _languages,
+  _srcRootPath,
+  _urls,
+  _urlSpec,
+  _wikiData,
+
+  _cachebust,
+  _developersComment,
+  _getSizeOfAdditionalFile,
+}) {
+  // Stub.
+  logInfo`So we are back in the mine!`;
+  logInfo`We are swinging our pickaxe in-`;
+  logInfo`...multiple directions,`;
+  logInfo`...multiple directions.`;
+  return true;
+}
